+++ pinot-core/src/main/java/org/apache/pinot/core/transport/QueryRouter.java
@@ -122,6 +122,9 @@ public class QueryRouter {
       ServerRoutingInstance serverRoutingInstance = entry.getKey();
       ServerChannels serverChannels = serverRoutingInstance.isTlsEnabled() ? _serverChannelsTls : _serverChannels;
       try {
+        // Record stats related to query submission just before sending the request. Otherwise, if the response is
+        // received immediately, there's a possibility of updating query response stats before updating query
+        // submission stats.
         _serverRoutingStatsManager.recordStatsAfterQuerySubmission(requestId, serverRoutingInstance.getInstanceId());
         serverChannels.sendRequest(rawTableName, asyncQueryResponse, serverRoutingInstance, entry.getValue(),
             timeoutMs);
@@ -183,7 +186,7 @@ public class QueryRouter {
 
       // Record query completion stats immediately after receiving the response from the server instead of waiting
       // for the reduce phase.
-      int latencyMs = asyncQueryResponse.getServerResponseDelayMs(serverRoutingInstance);
+      long latencyMs = asyncQueryResponse.getServerResponseDelayMs(serverRoutingInstance);
       _serverRoutingStatsManager.recordStatsUponResponseArrival(requestId, serverRoutingInstance.getInstanceId(),
           latencyMs);
     }
